El front esta enviando datos a la BD con la API de Laravel, falta validar los campos.
