+++ package.json
@@ -12,6 +12,7 @@
     "axios": "^1.3.3",
     "bootstrap": "^5.2.3",
     "bootstrap-icons": "^1.8.1",
+    "dotenv": "^16.0.3",
     "formik": "^2.2.9",
     "gh-pages": "^3.2.3",
     "lodash": "^4.17.21",
