+++ src/App.js
@@ -1,3 +1,4 @@
+import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import React from "react";
 import Menu from "./components/Menu";
 import Hero from "./components/Hero";
@@ -5,24 +6,29 @@ import Collections from "./components/Collections";
 import Discount from "./components/Discount";
 import HappyCustomer from "./components/HappyCustomer";
 import About from "./components/About";
-import CreateContact from "./components/CreateContact";
+import InsertContact from "./components/InsertContact";
 import Footer from "./components/Footer";
 
-
 export default function App() {
+  
   return (
-    <>
-    
-      <Menu />
-      <Hero />
-      <Collections />
-      <Discount />
-      <HappyCustomer />
-      <About />
-      <CreateContact/>
-      <Footer />
-      
-      
-    </>
+      <>
+
+
+          <Menu />
+          <Hero />
+          <Collections />
+          <Discount />
+          <HappyCustomer />
+          <About />
+          <div className="App">
+              <BrowserRouter>
+                  <Routes>
+                      <Route path="/" element={<InsertContact />} />
+                  </Routes>
+              </BrowserRouter>
+          </div>
+          <Footer />
+      </>
   );
 }
